MNT: move auxiliary functions from Function to tools.py
